完善功能，以及实现dao层

+++ src/main/java/com/nf/dao/AdministratorsDao.java
@@ -15,19 +15,19 @@ public interface AdministratorsDao {
      * @param administratorsEntity 注册管理员
      * @return 被影响的行
      */
-    Long signUp(AdministratorsEntity administratorsEntity);
+    int signUp(AdministratorsEntity administratorsEntity);
 
     /**
      * 修改账号
      * @param administratorsEntity 修改的账号
      * @return 被影响的行
      */
-    Long update(AdministratorsEntity administratorsEntity);
+    int update(AdministratorsEntity administratorsEntity);
 
     /**
      * 删除账号
      * @param id 被删除账号id
      * @return 被影响的行
      */
-    Long delete(int id);
+    int delete(int id);
 }
